requirements.txt
pynput==1.7.6
pyaudio==0.2.14
openai-whisper==20231117
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
